[client/api-client-extensions] move to ceremonies module
